refactor geoip code to reuse for mmdb
